Fix the issue where DummyProductModel was still used by some viewmodels
